Add concept of missing files for Text input. With these missing files comes error handling as well.
git-svn-id: svn://source.pentaho.org/svnkettleroot/mirror/Kettle/trunk@416 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ test/useCases/replay/be/ibridge/kettle/trans/step/KettleStepUseCase.java
@@ -33,18 +33,20 @@ public abstract class KettleStepUseCase extends TestCase {
 	public abstract String getFileExtension();
 
 	protected void tearDown() throws Exception {
-		File[] files = new File(directory).listFiles(new FilenameFilter() {
-			public boolean accept(File dir, String name) {
-				return !(name.endsWith("ktr")
-						|| name.endsWith(getFileExtension())
-						|| name.endsWith(".svn") || name
-						.endsWith(REPLAY_DATE + ".line"));
-			}
-		});
-		if (files != null)
-			for (int i = 0; i < files.length; i++) {
-				files[i].delete();
-			}
+		if (directory != null) {
+			File[] files = new File(directory).listFiles(new FilenameFilter() {
+				public boolean accept(File dir, String name) {
+					return !(name.endsWith("ktr")
+							|| name.endsWith(getFileExtension())
+							|| name.endsWith(".svn") || name
+							.endsWith(REPLAY_DATE + ".line"));
+				}
+			});
+			if (files != null)
+				for (int i = 0; i < files.length; i++) {
+					files[i].delete();
+				}
+		}
 		super.tearDown();
 	}
